moved compute to interface
