+++ responders/src/UI/embc-responder/src/app/app.module.ts
@@ -11,7 +11,7 @@ import { ErrorHandlingModule } from './shared/error-handling/error-handing.modul
 import { SharedModule } from './shared/shared.module';
 import { DatePipe } from '@angular/common';
 import { NgIdleKeepaliveModule } from '@ng-idle/keepalive';
-import { Compute } from './core/services/compute/compute';
+import { Compute } from './core/interfaces/compute';
 import { ComputeAppBaseService } from './core/services/compute/computeAppBase.service';
 import { ComputeFeaturesService } from './core/services/compute/computeFeatures.service';
 import { ComputeWizardService } from './core/services/compute/computeWizard.service';
