+++ responders/src/UI/embc-responder/src/app/core/interfaces/compute.ts
