+++ responders/src/UI/embc-responder/src/app/feature-components/search/evacuee-search/evacuee-search.component.html
@@ -5,10 +5,6 @@
         (showDataEntryComponent)="changeDataEntryComponent($event)"
         (showIDPhotoComponent)="changeVerifyIdComponent($event)"
       ></app-search-options>
-      <!-- <app-data-entry
-        (showDataEntryComponent)="changeDataEntryComponent($event)"
-        (showIDPhotoComponent)="changeVerifyIdComponent($event)"
-      ></app-data-entry> -->
     </ng-container>
 
     <ng-container *ngIf="showPhotoIDComponent">
