+++ responders/src/UI/embc-responder/src/app/feature-components/wizard/profile-components/profile-review/profile-review.component.ts
@@ -3,7 +3,6 @@ import {
   AbstractControl,
   FormBuilder,
   FormGroup,
-  ValidatorFn,
   Validators
 } from '@angular/forms';
 import { Router } from '@angular/router';
